requirements.txt
flask>=3.0.0
anthropic>=0.84.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
